Refactor notification service to use navigation factory
This commit refactors the notification service to use a navigation factory for navigating to different screens based on the notification type. The `NotificationRoutes` class has been updated to use the factory to get the appropriate navigation object based on the notification type. The `NotificationNavigationFactory` class has been added to map notification types to their corresponding navigation objects. The `NotificationType` enum has been moved to a separate file `navigation_types.dart`. This change improves the maintainability and extensibility of the notification service.